O21: rename to Oddities everywhere
