--- O21.MRB/O21.MRB.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\O21.StreamUtil\O21.StreamUtil.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Oxage.Wmf" Version="0.6.1" />
-    </ItemGroup>
-
-</Project>
