--- O21.NE/O21.NE.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <PropertyGroup>
-        <Description>New Executable (Win16 executable files, NE) file format reader.</Description>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\O21.StreamUtil\O21.StreamUtil.csproj" />
-    </ItemGroup>
-
-</Project>
