--- O21.Resources/O21.Resources.fsproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <Compile Include="Dib.fs" />
-        <Compile Include="Graphics.fs" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\O21.NE\O21.NE.csproj" />
-    </ItemGroup>
-
-</Project>
