--- O21.StreamUtil/O21.StreamUtil.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-    
-    <ItemGroup>
-        <InternalsVisibleTo Include="O21.MRB" />
-        <InternalsVisibleTo Include="O21.NE" />
-        <InternalsVisibleTo Include="O21.WinHelp" />
-    </ItemGroup>
-
-</Project>
