+++ Oddities.MRB/Oddities.MRB.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Oddities.StreamUtil\Oddities.StreamUtil.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Oxage.Wmf" Version="0.6.1" />
+    </ItemGroup>
+
+</Project>
