+++ Oddities.NE/Oddities.NE.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <PropertyGroup>
+        <Description>New Executable (Win16 executable files, NE) file format reader.</Description>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Oddities.StreamUtil\Oddities.StreamUtil.csproj" />
+    </ItemGroup>
+
+</Project>
