+++ Oddities.WinHelp/Topics/FormatHeader.cs
@@ -1,0 +1,19 @@
+using Oddities.StreamUtil;
+
+namespace Oddities.WinHelp.Topics;
+
+public struct FormatHeader
+{
+    public ushort FormatSize;
+    public byte Flags;
+    public ushort DataSize;
+
+    public static FormatHeader Read(Stream input)
+    {
+        FormatHeader header;
+        header.FormatSize = input.ReadCompressedUInt16();
+        header.Flags = input.ReadByteExact();
+        header.DataSize = input.ReadCompressedUInt16();
+        return header;
+    }
+}
