+++ Oddities.WinHelp/Topics/TopicBlockHeader.cs
@@ -1,0 +1,19 @@
+﻿using Oddities.StreamUtil;
+
+namespace Oddities.WinHelp.Topics;
+
+public struct TopicBlockHeader
+{
+    public int LastParagraph;
+    public int TopicData;
+    public int LastTopicHeader;
+
+    public static TopicBlockHeader Load(Stream input)
+    {
+        TopicBlockHeader header;
+        header.LastParagraph = input.ReadInt32Le();
+        header.TopicData = input.ReadInt32Le();
+        header.LastTopicHeader = input.ReadInt32Le();
+        return header;
+    }
+}
