+++ Oddities.sln
@@ -20,6 +20,16 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "workflows", "workflows", "{
 		.github\workflows\main.yml = .github\workflows\main.yml
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Oddities.MRB", "Oddities.MRB\Oddities.MRB.csproj", "{4FDF05F5-61B1-41EF-884F-CB2A86855FC1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Oddities.NE", "Oddities.NE\Oddities.NE.csproj", "{04401361-D9C4-44BA-A0AD-E187C398A7D5}"
+EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Oddities.Resources", "Oddities.Resources\Oddities.Resources.fsproj", "{6C005755-1B20-4E20-8DAB-462DAD4FEA0A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Oddities.StreamUtil", "Oddities.StreamUtil\Oddities.StreamUtil.csproj", "{BF1E8FDC-6EF8-47AA-B2E1-DBFBDEB0ABC0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Oddities.WinHelp", "Oddities.WinHelp\Oddities.WinHelp.csproj", "{7878A7E9-6CD0-4180-BB33-DD64E45F78F6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +40,27 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{12CFB952-1E87-4BDD-9EB1-E3005CA6B0BC} = {EF058526-C7EE-49BB-9104-3459B7BAA839}
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4FDF05F5-61B1-41EF-884F-CB2A86855FC1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4FDF05F5-61B1-41EF-884F-CB2A86855FC1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4FDF05F5-61B1-41EF-884F-CB2A86855FC1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4FDF05F5-61B1-41EF-884F-CB2A86855FC1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{04401361-D9C4-44BA-A0AD-E187C398A7D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{04401361-D9C4-44BA-A0AD-E187C398A7D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{04401361-D9C4-44BA-A0AD-E187C398A7D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{04401361-D9C4-44BA-A0AD-E187C398A7D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6C005755-1B20-4E20-8DAB-462DAD4FEA0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6C005755-1B20-4E20-8DAB-462DAD4FEA0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6C005755-1B20-4E20-8DAB-462DAD4FEA0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6C005755-1B20-4E20-8DAB-462DAD4FEA0A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BF1E8FDC-6EF8-47AA-B2E1-DBFBDEB0ABC0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BF1E8FDC-6EF8-47AA-B2E1-DBFBDEB0ABC0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BF1E8FDC-6EF8-47AA-B2E1-DBFBDEB0ABC0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BF1E8FDC-6EF8-47AA-B2E1-DBFBDEB0ABC0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7878A7E9-6CD0-4180-BB33-DD64E45F78F6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7878A7E9-6CD0-4180-BB33-DD64E45F78F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7878A7E9-6CD0-4180-BB33-DD64E45F78F6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7878A7E9-6CD0-4180-BB33-DD64E45F78F6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
